+++ frontend/src/components/host/JobListCard/JobBox/index.tsx
