+++ frontend/src/components/host/JobListCard/JobBox/styles.ts
