+++ frontend/src/components/host/JobListCard/JobBox/useJobBox.tsx
