+++ frontend/src/components/host/SpaceInfo/SpaceInfoCreateBox/index.tsx
