+++ frontend/src/components/host/SpaceInfo/SpaceInfoCreateBox/styles.ts
