+++ frontend/src/components/host/SpaceInfo/SpaceInfoCreateBox/useSpaceCreateForm.ts
