+++ frontend/src/components/host/SpaceInfo/SpaceInfoDisplayBox/index.tsx
