+++ frontend/src/components/host/SpaceInfo/SpaceInfoDisplayBox/styles.ts
