+++ frontend/src/components/host/SpaceInfo/SpaceInfoDisplayBox/useSpaceInfoDisplayBox.ts
