+++ frontend/src/components/host/SpaceInfo/SpaceInfoUpdateBox/index.tsx
