+++ frontend/src/components/host/SpaceInfo/SpaceInfoUpdateBox/styles.ts
