+++ frontend/src/components/user/SectionCard/SectionInfoPreview/index.tsx
