+++ frontend/src/components/user/SectionCard/SectionInfoPreview/styles.ts
