+++ frontend/src/pages/host/DashBoard/styles.ts
@@ -116,6 +116,36 @@ const buttons = css`
   }
 `;
 
-const styles = { layout, contents, cardWrapper, slackButton, linkButton, buttons };
+const spaceDeleteButton = css`
+  width: auto;
+  height: 2rem;
+  padding: 0 12px;
+  font-size: 0.9rem;
+  margin: 0 0 16px 12px;
+  background-color: ${theme.colors.red};
+  color: ${theme.colors.white};
+
+  svg {
+    height: 24px;
+    width: 24px;
+    transform: translateY(3px);
+  }
+
+  @media screen and (min-width: 1024px) {
+    svg {
+      height: 16px;
+      width: 16px;
+      margin-right: 4px;
+    }
+  }
+
+  @media screen and (max-width: 1023px) {
+    span {
+      display: none;
+    }
+  }
+`;
+
+const styles = { layout, contents, cardWrapper, slackButton, linkButton, buttons, spaceDeleteButton };
 
 export default styles;
