+++ frontend/src/pages/host/DashBoard/useDashBoard.tsx
@@ -1,7 +1,8 @@
-import { useQuery } from 'react-query';
-import { useParams } from 'react-router-dom';
+import { useMutation, useQuery } from 'react-query';
+import { useNavigate, useParams } from 'react-router-dom';
 
 import SlackUrlModal from '@/components/host/SlackUrlModal';
+import SpaceDeleteModal from '@/components/host/SpaceDeleteModal';
 
 import useModal from '@/hooks/useModal';
 import useToast from '@/hooks/useToast';
@@ -14,7 +15,9 @@ import apiSubmission from '@/apis/submission';
 import { ID } from '@/types';
 
 const useDashBoard = () => {
-  const { openModal } = useModal();
+  const navigate = useNavigate();
+
+  const { openModal, closeModal } = useModal();
   const { openToast } = useToast();
 
   const { spaceId } = useParams() as { spaceId: ID };
@@ -44,12 +47,41 @@ const useDashBoard = () => {
     },
   });
 
+  const { refetch } = useQuery(['deleteSpaces'], apiSpace.getSpaces, {
+    enabled: false,
+    onSuccess: data => {
+      const { spaces } = data;
+
+      if (spaces.length === 0) {
+        navigate('/host/manage/spaceCreate');
+        return;
+      }
+
+      const space = spaces[0];
+      navigate(`/host/manage/${space.id}`);
+    },
+  });
+
+  const { mutate: deleteSpace } = useMutation((spaceId: ID) => apiSpace.deleteSpace(spaceId), {
+    onSuccess: () => {
+      refetch();
+      closeModal();
+      openToast('SUCCESS', '공간이 삭제되었습니다.');
+    },
+  });
+
   const onClickSlackButton = () => {
     openModal(<SlackUrlModal jobs={jobsData?.jobs || []} />);
   };
 
   const onClickLinkButton = () => {
     copyEntranceLink();
+  };
+
+  const onClickDeleteSpace = () => {
+    openModal(
+      <SpaceDeleteModal text={`${spaceData?.name} 공간을 삭제합니다` || ''} onClick={() => deleteSpace(spaceId)} />
+    );
   };
 
   return {
@@ -59,6 +91,7 @@ const useDashBoard = () => {
     submissionData,
     onClickSlackButton,
     onClickLinkButton,
+    onClickDeleteSpace,
   };
 };
